use prepared statements
We are not using prepared statements in this code, and as a consequence
it is blowing up the prepared statement cache in the scylla side with
negative consequences to other applications running in the same cluster.

Using CQL tracing, we can see that before each query a prepare attempt
is made to Scylla.

This patch adds a prepare step and improves this behavior.

Signed-off-by: Glauber Costa <glauber@scylladb.com>