operator: demo added
It has an interactive script that allows the presenter to hold and
continue easy with a simple key stroke.

The demo shows deployment, scaling up, upgrade scylla version and
load testing using c-stress. Monitoring is deployed and is used to
show the load as well as the scaling of the node count.