Archive ScyllaDB Helm Chart
This Helm chart is not supported nor tested, but it shows up higher in
Google search results than maintained and supported Scylla Operator
which also has a Helm Chart.
At least two users tried to use it and failed to create Scylla Cluster,
so it's better to point them to Helm Chart we support.

Previous code examples will be avaialbe in this repository Git history.