Rust_Scylla_Driver/chat: use chrono DateTime instead of i64
Since version 0.11 for serialization and 0.15 for deserialization Rust
Driver verifies the provided types against prepared statement metadata.
That means it is no longer possible to use "i64" to serialize /
deserialize CQL timestamps. User needs to use either our thing wrapper
(CqlTimestamp type) or types from chrono / time libraries.

chrono was already a dependency in this example, so I decided to migrate
to it.